Cloud-Automation/GCP/LAB03-IAM-Service-Account-Creation/requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
python-dotenv==1.0.0 